requirements.txt
# Core ML libraries
tensorflow>=2.12.0
keras>=2.12.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
pydicom>=2.4.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
kaggle>=1.5.16
PyYAML>=6.0
h5py>=3.9.0
tqdm>=4.65.0

# Medical imaging specific
SimpleITK>=2.2.1
nibabel>=5.1.0
medpy>=0.4.0

# Visualization and analysis
plotly>=5.15.0
tensorboard>=2.12.0
grad-cam>=1.4.6

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.7.0
flake8>=6.0.0
pytest>=7.4.0

# Performance
albumentations>=1.3.1
imgaug>=0.4.0

# MLOps
mlflow>=2.5.0
wandb>=0.15.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
requests>=2.31.0
